// src/app/models/userCompanyDTO.ts
export class UserCompanyDTO {

    id: number;
    companyOwner: string;
    legalBusinessName: string;
    zipCode: string;
    phoneNo: string;
    companyEmail: string;
    website: string;
    licenseNo: string;
    taxId: string;
    blobImageURI: string;
    imageName: string;
    companyGUID: string;
    cityId: number;
    cityName: string;
    provinceName: string;
    countryName: string;
    countryId: number;
    provinceId: number;

    ownerName: string;
    ownerCountryId: number;
    ownerPhoneNo: string;
    primaryOwnerEmail: string;
    ownerMembershipTypeId: number;
    businessStructureId: number;
    ownerLicenseNo: string;
}